+++ src/constants.ts
@@ -1,3 +1,3 @@
 export const APP_VERSION = __APP_VERSION__;
-export const REPOSITORY_URL = "https://github.com/XxMinor/mykvm";
+export const REPOSITORY_URL = "https://github.com/yunbim/mykvm";
 export const RELEASES_URL = `${REPOSITORY_URL}/releases/latest`;